isl_tab_basic_set_constrained_lexmin: allow imposing fixed values
That is, allow the caller to force linear combinations of
the schedule coefficients to be equal to some fixed values.
This can be done in addition to the non-zero constraints or
as an alternative to the non-zero constraints.
If a fixed value is combined with a non-zero constraint, then
the fixed value is imposed in all of the non-zero cases.
If a fixed value is imposed separately, then a single case
is considered in which the constraint is imposed.

Note that support for imposing fixed values in
isl_tab_basic_set_constrained_lexmin is not very useful on its own.
The caller could simply impose the constraints directly in the ILP
problem.  This support will become useful, however, when
the constraints can be made optional.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>