add isl_schedule_constraints_set_inter_consecutivity
These constraints allow the user to specify that some pairs
of statement instances should be executed consecutively
at some level of the schedule.
In particular, each constraint in the list relates pairs
of instances of two distinct statements and references
a pair of intra-statement consecutivity constraints.
The scheduler will try to make those pairs of instances
execute consecutively at the schedule level that corresponds
to the first rows of the inner parts of
the intra-statement consecutivity constraints.

The typical use case it to extend the consecutive accesses
within a pair of statements as specified by
the intra-statement consecutivity constraints to also
be consecutive across those two statement.

In contrast to most other schedule constraints,
the inter-statement consecutivity constraints are ordered.
This allows the user to specify the order in which
the constraints should be considered, for example
to achieve inter-statement consecutivity at the innermost
level first.

The representation is a bit involved with the embedded
references to a pair of intra-statement consecutivity constraints.
Other representations that were considered tried to combine
all information in a single isl_pw_multi_aff object.
e.g., by having a wrapped copy of the relation as the domain
of the isl_pw_multi_aff object and the product of the two
intra-statement consecutivity constraints as associated function.
However, this would allow equality constraints in the relation
to be used to simplify the function, potentially modifying
the intra-statement consecutivity constraints.
An alternative would be for the domain to be a product too,
once to specify the relation and once to serve as a (universe)
domain of the function, but this is rather involved as well and
results in function being defined in terms of only one of the four
spaces in the domain.

The constraints are enforced by imposing a schedule distance
of zero while the outer parts of the intra-statement
consecutivity constraints are being covered and
a schedule distance of one when the first row of the inner parts
of both intra-statement consecutivity constraints is considered.

Since it is not immediately obvious how to translate
inter-statement consecutivity constraints to constraints
between clusters of statements during the incremental scheduling,
the inter-statement consecutivity constraints are always handled
in terms of the original statements.  This means that any pair
of clusters that is connected through one or more
inter-statement consecutivity constraints is combined into a single
cluster.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>