introduce (unchecked) cpp bindings generation
The design of these bindings is essentially the same as
that of the checked C++ bindings, except that
the user does not need to explicitly check for errors.
In particular, no isl::boolean or isl::stat needs to be introduced.
Instead, exceptional results (negative isl_bool and isl_stat,
but also NULL pointer results) cause an exception to be thrown,
if exceptions are available.

If exceptions are not available then isl aborts on any error condition.
Note that exceptions are a standard feature of C++, but some
compilers allow the user to turn off support for exceptions.
Unfortunately, it does not appear to be possible to determine
that this feature was removed, but there are some vendor specific ways
for checking that exceptions are available.
This means the bindings need to assume that exceptions are not available
unless it can explicitly determine that exceptions are available.

Exceptions thrown by C++ callbacks are caught and rethrown,
just like in the Python bindings.

Additionally, all arguments of methods are checked for validity.
This ensures that a valid isl_ctx is available when
an exception needs to be thrown based on an error that occurred
during the execution of the C function.
The exception class derives from std::exception rather than std::logic_error
or std::runtime_error because only some types of isl exceptions
should be considered as logic errors, while some other can
be considered as runtime errors.

The header file is created and installed as isl/cpp.h

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>