isl_scheduler.c: isl_sched_node: replace sched_map by band_sched
The field sched_map is the value cached by node_extract_schedule,
which is used to restrict dependences to those mapped to the same
value by the current schedule.  However, after the construction
of a band is completed, all dependence relations are already
updated to pairs with equal schedule values at that point.
During the construction of a band, only the effect of the current band
therefore needs to be taken into account.
Replace node_extract_schedule by node_extract_band_schedule and
change the isl_sched_node field accordingly.

This change removes the final use of the current schedule as a whole
for purposes other than avoiding linear dependence of schedule rows.
This will make it possible to include the information derived
from the prefix schedule constraint without having to worry that
it might have any effects beyond the linear dependence avoidance.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>